+++ api/rust/include/LIEF/rust/DWARF/types/Enum.hpp
@@ -15,10 +15,40 @@
 #pragma once
 #include "LIEF/DWARF/types/Enum.hpp"
 #include "LIEF/rust/DWARF/Type.hpp"
+#include "LIEF/rust/optional.hpp"
+#include "LIEF/rust/Iterator.hpp"
+
+class DWARF_types_Enum_Entry : public Mirror<LIEF::dwarf::types::Enum::Entry> {
+  public:
+  using Mirror::Mirror;
+  using lief_t = LIEF::dwarf::types::Enum::Entry;
+
+  auto name() const { return get().name(); }
+
+  int64_t value(uint32_t& is_set) const {
+    return details::make_optional(get().value(), is_set);
+  }
+};
 
 class DWARF_types_Enum : public DWARF_Type {
   public:
   using lief_t = LIEF::dwarf::types::Enum;
+
+  class it_entries :
+      public ContainerIterator<DWARF_types_Enum_Entry,
+                               std::vector<LIEF::dwarf::types::Enum::Entry>>
+  {
+    public:
+    using container_t = std::vector<LIEF::dwarf::types::Enum::Entry>;
+    it_entries(container_t content)
+      : ContainerIterator(std::move(content)) { }
+    auto next() { return ContainerIterator::next(); }
+  };
+
+  auto entries() const {
+    std::vector<LIEF::dwarf::types::Enum::Entry> entries = impl().entries();
+    return std::make_unique<it_entries>(std::move(entries));
+  }
 
   static bool classof(const DWARF_Type& type) {
     return lief_t::classof(&type.get());