+++ doc/sphinx/changelog.rst
@@ -12,6 +12,7 @@
 
 :DWARF:
 
+  * Add support to access Enum entries: |lief-dwarf-types-Enum-entries|
   * Add support to read or assign a register to a function's parameter
   * Add support to read or assign a description (``DW_AT_description``) to
     a |lief-dwarf-function|, |lief-dwarf-variable|, |lief-dwarf-lexical-block|: